refactor(python): extract the NMF + LSA model legs (python/nmf_lsa.rs)
Fourth step of the python.rs decomposition — the first *model* legs, proving the
per-model recipe the remaining families will follow. The NMF and LSA pyclasses
(+ their State structs and save/load impls) move to src/python/nmf_lsa.rs.

The recipe, now locked in: a model-leg file does `use super::*;` (the canonical
child-module pattern reaches python/mod.rs's private helpers — Corpus,
build_corpus_from_docs, run_with_threads, the array/error/save items, the pyo3/
numpy/crate imports), so the moved code is unchanged; mod.rs declares `mod
nmf_lsa;` and `use nmf_lsa::{LSA, NMF};` for #[pymodule] registration. No new
pub(crate) plumbing was needed beyond what the corpus PR already established.

Pure move, no behavior change: clippy -D warnings + fmt clean, cargo test
--workspace, maturin --features python, NMF/LSA fit+save/load smoke, pytest
(2782 passed, 26 skipped). faSTM untouched. mod.rs is ~650 lines lighter.
